==> proyecto integrador/1.CODIGO/1.tratamiento.ipynb <==
ARCHIVO DONDE SE TRATARA EL DATASET

ADEMAS SE PUEDEN AGREGAR A PACIENTES

1) IMPORTAMOS LIBRERIAS


# archivos csv
import pandas as pd


2. CON LA LIBRERIA PANDAS LEEMOS EL ARCHIVO CSV

df = pd.read_csv("0.dataset_diabetes.csv")



3.TRADUCIMOS LAS COLUMNAS DEL DATASET

df.columns = ["Genero", "Edad", "Hipertencion", "Cardiopatia", "Historial_fumador", "IMC", "Nivel_HnA1c", "Glucosa_en_sangre", "Diabetes"]
df

4.SE ELIMINARAN TODAS LAS FILAS QUE CONTENGAN LA INFORMACION "NO INFO" DEL "HISTORIAL DE FUMADOR"

df = df[df["Historial_fumador"] != "No Info"]



5.ELIMINAMOS EL INDICE ANTERIOR Y CREAMOS UNO DE FORMA ASCENDENTE

df = df.reset_index(drop=True)



6.TRADUCCION DE LOS STRINGS DEL DATASET

df["Genero"] = df["Genero"].replace({"Male": "Masculino", "Female": "Femenino"})
df["Historial_fumador"] = df["Historial_fumador"].replace({"never": "Nunca", "current": "Actualmente", "former": "Anteriormente", "ever": "Alguna vez", "not current": "Anteriormente"})


  7. AGREGAR PACIENTE :
MEDIANTE LA INTERACCION CON EL USUARIO, ESTE PODRA AGREGAR A PACIENTES.


# Cargar el dataset "df"
df = df

# obtener edad del usuario.
genero = int(input("Ingrese el género del paciente (0 para Masculino, 1 para Femenino): "))
# si es masculino ==0
if genero == 0:
    
    genero = "Masculino"
# si es femenino
elif genero == 1:
    genero = "Femenino"
# OBETNER EDAD DEL USUARIO
edad = int(input("Ingrese la edad del paciente: "))

hipertension = int(input("Ingrese si el paciente tiene hipertensión (0 para no, 1 para sí): "))
cardiopatia = int(input("Ingrese si el paciente tiene cardiopatía (0 para no, 1 para sí): "))

historial_fumador = int(input("Ingrese el historial de fumador del paciente (1 para Actualmente, 2 para Anteriormente, 3 para Nunca): "))
if historial_fumador == 1:
    historial_fumador = "Actualmente"
elif historial_fumador == 2:
    historial_fumador = "Anteriormente"
elif historial_fumador == 3:
    historial_fumador = "Nunca"

imc = int(input("Ingrese el índice de masa corporal (IMC) del paciente: "))
nivel_hba1c = int(input("Ingrese el nivel de HbA1c del paciente: "))
glucosa_sangre = int(input("Ingrese el nivel de glucosa en sangre del paciente: "))

# Crear una nueva fila con los datos ingresados
nueva_fila = pd.DataFrame([[genero, edad, hipertension, cardiopatia, historial_fumador, imc, nivel_hba1c, glucosa_sangre, 0]], columns=df.columns)

# Agregar la nueva fila al dataset existente
df = pd.concat([nueva_fila, df], ignore_index=True)




8.CLASIFICADOR: se clasificara a los pacientes segun su edad, y mostrara la cantidad de pacientes con o sin diabetes segun rango etario



# DefiNINIENDO LOS RANGOS DE EDAD:
# Rangos de edad: JOVENES[0, 30), ADULTOS[30, 60), PERSONAS MAYORES[60, 150)
edades = [0, 30, 60, 150] 

# Crea una nueva columna de grupo de edad en el dataset
df['Grupo de Edad'] = pd.cut(df['Edad'], bins=edades, labels=['Jóvenes', 'Adultos', 'Personas Mayores'])

# Calcula la distribución de la diabetes en cada grupo de edad
distribucion = df.groupby('Grupo de Edad')['Diabetes'].value_counts().unstack()




10.CLASIFICADOR
: SE CREARA UNA NUEVA COLUMNA "IMC CATEGORIA", LA CUAL ASIGNARA EN QUE CATEGRIA ESTAS SEGUN TU IMC, ESTO SE VERA RELACIONADO CON LAS PERSONAS QUE TIENEN DIABETES O NO




# Rangos de IMC

# Rangos de IMC: Bajo Peso [0, 18.5), Normal[18.5, 25), Sobrepeso[25, 30), Obeso[30, 100)

imc_ranges = [0, 18.5, 25, 30, 100]  

# Crea una nueva columna de categoría de IMC en el dataset
df['IMC Categoría'] = pd.cut(df['IMC'], bins=imc_ranges, labels=['Bajo Peso', 'Normal', 'Sobrepeso', 'Obeso'])

# Calcula la distribución de la diabetes en cada categoría de IMC
distribucion_imc = df.groupby('IMC Categoría')['Diabetes'].value_counts().unstack()




¡¡IMPORTANDO "DF" A INTERFAZ¡¡

import pickle

# Supongamos que `df` contiene tu dataset


# Guardar `df` en un archivo
with open('datos.pickle', 'wb') as archivo:
    pickle.dump(df, archivo)


¡¡IMPORTANDO "DF" A ALGORITMO LINEAL""

# Supongamos que `df` contiene tu dataset


# Guardar `df` en un archivo
with open('datos.pickle', 'wb') as algoritmo_lineal:
    pickle.dump(df, algoritmo_lineal)


¡¡IMPORTANDO "DF" A ALGORITMO BINARIO!!

# Supongamos que `df` contiene tu dataset


# Guardar `df` en un archivo
with open('datos.pickle', 'wb') as algoritmo_binario:
    pickle.dump(df, algoritmo_binario)

¡¡IMPORTANDO "DF" A GRAFICOS!!

# Guardar `df` en un archivo
with open('datos.pickle', 'wb') as graficos:
    pickle.dump(df, graficos)

df
