==> proyecto integrador/1.CODIGO/3.binario.ipynb <==
Se importan las bibliotecas necesarias.
Se carga el dataset para poder hacer un tratamiento de datos.


# PANDAS
import pandas as pd

# IMPORTAR DF
import pickle

# LIBRERIA PARA BUSQQUEDA LINEAL
from binarytree import build


¡IMPORTANDO DF!


# Cargar `df` desde el archivo

with open('datos.pickle', 'rb') as algoritmo_binario:
    df = pickle.load(algoritmo_binario)



Se implementa un algoritmo de busqueda binaria que contiene nodos y arboles.


index_dict = dict(zip(range(len(df)), df.index))

class Node:
    def __init__(self, valor):
        self.valor = valor
        self.izquierda = None
        self.derecha = None

def constructor(llaves):
    if not llaves:
        return None
    medio = len(llaves)//2

    nodo = Node(llaves[medio])
    nodo.izquierda = constructor(llaves[:medio])
    nodo.derecha = constructor(llaves[medio+1:])
    return nodo

indices = list(index_dict.keys())
arbol = constructor(indices)

def buscar_persona(indice, nodo):
    if nodo is None or nodo.valor == indice:
        return nodo
    if indice < nodo.valor:
        return buscar_persona(indice, nodo.izquierda)
    else:
        return buscar_persona(indice, nodo.derecha)

indice_busqueda = int(input("Ingrese el indice del paciente a buscar: "))

nodo_persona = buscar_persona(indice_busqueda, arbol)

if nodo_persona is not None:
    indice_persona =nodo_persona.valor
    resultado = df.loc[index_dict[indice_persona]]
    print(resultado)
else:
    print("No se encontro a ninguna persona con ese indice: ", indice_busqueda)

# Lógica de recomendación de tratamiento
# Tratamiento para la diabetes
if resultado['Diabetes'] == 0:
    print("Recomendación de tratamiento para la diabetes: Estilo de vida saludable y seguimiento médico regular.")
else:
    print("Recomendación de tratamiento para la diabetes: Control de la glucosa en sangre, medicación y seguimiento médico especializado.")

# Tratamiento para la hipertensión
if resultado['Hipertencion'] == 1:
    print("Recomendación de tratamiento para la hipertensión: Medicación para controlar la presión arterial y seguimiento médico regular.")

# Tratamiento para el historial de fumador
historial_fumador = resultado['Historial_fumador']
if historial_fumador == 1:
    print("Recomendación de tratamiento para el historial de fumador: Asesoramiento para dejar de fumar y seguimiento médico para evaluar los efectos del tabaquismo.")
elif historial_fumador == 2:
    print("Recomendación de tratamiento para el historial de fumador: Apoyo y recursos para evitar recaídas y seguimiento médico para evaluar los efectos del tabaquismo anterior.")
elif historial_fumador == 3:
    print("Recomendación de tratamiento para el historial de fumador: Evaluación de los efectos del tabaquismo pasado y asesoramiento para evitar futuras recaídas.")
elif historial_fumador == 4:
    print("Recomendación de tratamiento para el historial de fumador: Promoción de hábitos saludables y seguimiento médico para prevenir el inicio del tabaquismo.")

# Tratamiento para el IMC
imc = resultado['IMC']
if imc >= 30:
    print("Recomendación de tratamiento para la obesidad: Cambios en la alimentación, actividad física regular y seguimiento médico para controlar el peso.")
elif imc >= 25:
    print("Recomendación de tratamiento para el sobrepeso: Modificaciones en la dieta, aumento de la actividad física y seguimiento médico para controlar el peso.")

# Lógica de recomendación de tratamiento
# Tratamiento para la diabetes
if resultado['Diabetes'] == 0:
    print("Recomendacion 1 por si no  tiene Diabetes.")
else:
    print("Recomendacion 2 por si es Diabetico")

# Tratamiento para la hipertensión
if resultado['Hipertencion'] == 1:
    print("Recomendación de tratamiento por si tiene Hipertencion.")

# Tratamiento para el historial de fumador
historial_fumador = resultado['Historial_fumador']
if historial_fumador == 1:
    print("Recomendación por si fuma Actualmente.")
elif historial_fumador == 2:
    print("Recomendación por si el Paciente Fumaba anteriormente.")
elif historial_fumador == 3:
    print("Recomendación por si nunca ha fumado")
elif historial_fumador == 4:
    print("Recomendación por si alguna vez fumo.")

# Tratamiento para el IMC
imc = resultado['IMC']
if imc >= 30:
    print("Recomendación de tratamiento para la obesidad")
elif imc >= 25:
    print("Recomendación de tratamiento para el sobrepeso")

df
